+++ anni-repo/src/models/album.rs
@@ -184,6 +184,9 @@ impl Album {
         toml::to_string_pretty(&self).unwrap()
     }
 
+    /// Apply album metadata to a directory formatted with strict album format.
+    ///
+    /// This function applies both metadata and cover.
     #[cfg(feature = "apply")]
     pub fn apply_strict<P>(&self, directory: P) -> Result<(), crate::error::AlbumApplyError>
     where
@@ -228,11 +231,12 @@ impl Album {
         }
 
         for (index, disc_id) in discs.iter().enumerate() {
+            let disc_path = directory.as_ref().join(disc_id);
             if disc_id != &(index + 1).to_string() {
-                return Err(AlbumApplyError::InvalidDiscFolder);
+                return Err(AlbumApplyError::InvalidDiscFolder(disc_path));
             }
 
-            let disc_cover_path = directory.as_ref().join(disc_id).join("cover.jpg");
+            let disc_cover_path = disc_path.join("cover.jpg");
             if !disc_cover_path.exists() {
                 return Err(AlbumApplyError::MissingCover(disc_cover_path));
             }
@@ -309,6 +313,95 @@ impl Album {
                 // if modified {
                 flac.save::<String>(None)?;
                 // }
+            }
+        }
+        Ok(())
+    }
+
+    /// Apply album metadata to a directory formatted with **convention album format**.
+    ///
+    /// This function applies metadata only. Cover is not checked
+    #[cfg(feature = "apply")]
+    pub fn apply_convention<P>(&self, directory: P) -> Result<(), crate::error::AlbumApplyError>
+    where
+        P: AsRef<Path>,
+    {
+        use crate::error::AlbumApplyError;
+        use anni_common::fs;
+        use anni_flac::{
+            blocks::{UserComment, UserCommentExt},
+            FlacHeader,
+        };
+
+        let disc_total = self.discs_len();
+
+        for (disc_num, disc) in self.iter().enumerate() {
+            let disc_num = disc_num + 1;
+            let disc_dir = if disc_total > 1 {
+                directory.as_ref().join(format!(
+                    "[{catalog}] {title} [Disc {disc_num}]",
+                    catalog = disc.catalog(),
+                    title = disc.title(),
+                    disc_num = disc_num,
+                ))
+            } else {
+                directory.as_ref().to_owned()
+            };
+
+            if !disc_dir.exists() {
+                return Err(AlbumApplyError::InvalidDiscFolder(disc_dir));
+            }
+
+            let files = fs::get_ext_files(&disc_dir, "flac", false)?;
+            let tracks = disc.iter();
+            let track_total = disc.tracks_len();
+            if files.len() != track_total {
+                return Err(AlbumApplyError::TrackMismatch {
+                    path: disc_dir,
+                    expected: track_total,
+                    actual: files.len(),
+                });
+            }
+
+            for (track_num, (file, track)) in files.iter().zip(tracks).enumerate() {
+                let track_num = track_num + 1;
+
+                let mut flac = FlacHeader::from_file(file)?;
+                let comments = flac.comments();
+                // TODO: read anni convention config here
+                let meta = format!(
+                    r#"TITLE={title}
+ALBUM={album}
+ARTIST={artist}
+DATE={release_date}
+TRACKNUMBER={track_number}
+TRACKTOTAL={track_total}
+DISCNUMBER={disc_number}
+DISCTOTAL={disc_total}
+"#,
+                    title = track.title(),
+                    album = disc.title(),
+                    artist = track.artist(),
+                    release_date = self.release_date(),
+                    track_number = track_num,
+                    track_total = track_total,
+                    disc_number = disc_num,
+                    disc_total = disc_total,
+                );
+                // no comment block exist, or comments is not correct
+                if comments.is_none() || comments.unwrap().to_string() != meta {
+                    let comments = flac.comments_mut();
+                    comments.clear();
+                    comments.push(UserComment::title(track.title()));
+                    comments.push(UserComment::album(disc.title()));
+                    comments.push(UserComment::artist(track.artist()));
+                    comments.push(UserComment::date(self.release_date()));
+                    comments.push(UserComment::track_number(track_num));
+                    comments.push(UserComment::track_total(track_total));
+                    comments.push(UserComment::disc_number(disc_num));
+                    comments.push(UserComment::disc_total(disc_total));
+                    flac.save::<String>(None)?;
+                }
             }
         }
         Ok(())